flask/requirements.txt
Flask
Gunicorn
pillow
torch
torchvision